agentic_assistant/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
python-multipart==0.0.20
sentence-transformers==3.4.1
hnswlib==0.8.0
numpy==2.2.3
pypdf==5.3.1
python-dotenv==1.0.1
httpx==0.28.1
requests==2.32.3
sqlite-utils==3.36
Markdown==3.6
groq==0.13.1
google-generativeai==0.8.4
openai==1.61.1
